refactor: 패키지 생성 - static class를 패키지 안에 class 로 생성
