+++ src/main/java/onboarding/Problem2.java
@@ -1,50 +1,8 @@
 package onboarding;
 
-import java.util.Stack;
-import java.util.stream.Collectors;
+import onboarding.problem2.Decryption;
 
 public class Problem2 {
-  static class Decryption {
-    private final Stack<Character> message = new Stack<>();
-    private Character duplicateCode;
-
-    public void decrypt(Character code) {
-      if (isDuplicateCode(code)) {
-        message.pop();
-        saveDuplicateCode(code);
-        return;
-      }
-      if (isSameAsDuplicateCode(code)) {
-        return;
-      }
-      message.push(code);
-      initializeDuplicateCode();
-    }
-
-    private void saveDuplicateCode(Character code) {
-      duplicateCode = code;
-    }
-
-    private void initializeDuplicateCode() {
-      duplicateCode = null;
-    }
-
-    private boolean isSameAsDuplicateCode(Character code) {
-      return this.duplicateCode == code;
-    }
-
-
-    private boolean isDuplicateCode(Character code) {
-      return !message.empty() && message.peek() == code;
-    }
-
-    public String getMessage() {
-      return message.stream()
-              .map(Object::toString)
-              .collect(Collectors.joining());
-    }
-  }
-
   public static String solution(String cryptogram) {
     Decryption decryption = new Decryption();
     cryptogram.chars()
